chore: moved utils to client folder
