+++ client/components/speech-wpm-d3.js
@@ -1,5 +1,5 @@
 import * as d3 from 'd3'
-import {formatSeconds} from '../../utils'
+import {formatSeconds} from '../utils'
 
 const MARGIN = {TOP: 10, BOTTOM: 50, LEFT: 70, RIGHT: 10}
 const WIDTH = 650 - MARGIN.LEFT - MARGIN.RIGHT
